Use 64-bit type to represent microseconds.
int and long are 32-bit on Windows, causing overflows.

(cherry picked from commit 85b19784086faa7676d030f03bfc4801705c180f)

+++ 3d-viewer/3d_canvas/create_layer_items.cpp
@@ -154,9 +154,9 @@ void BOARD_ADAPTER::createLayers( REPORTER* aStatusReporter )
     //    https://github.com/KiCad/kicad-source-mirror/blob/master/3d-viewer/3d_draw.cpp#L692
 
 #ifdef PRINT_STATISTICS_3D_VIEWER
-    unsigned stats_startCopperLayersTime = GetRunningMicroSecs();
+    int64_t stats_startCopperLayersTime = GetRunningMicroSecs();
 
-    unsigned start_Time = stats_startCopperLayersTime;
+    int64_t start_Time = stats_startCopperLayersTime;
 #endif
 
     PCB_LAYER_ID cu_seq[MAX_CU_LAYERS];
